+++ src/Core/R4300/x86/GRegImm.cpp
@@ -50,12 +50,12 @@ void genbltz_test()
 void genbltz()
 {
 #ifdef INTERPRET_BLTZ
-    gencallinterp((uint32_t)BLTZ, 1);
+    gencallinterp((uintptr_t)BLTZ, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZ, 1);
+        gencallinterp((uintptr_t)BLTZ, 1);
         return;
     }
 
@@ -68,12 +68,12 @@ void genbltz()
 void genbltz_out()
 {
 #ifdef INTERPRET_BLTZ_OUT
-    gencallinterp((uint32_t)BLTZ_OUT, 1);
+    gencallinterp((uintptr_t)BLTZ_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZ_OUT, 1);
+        gencallinterp((uintptr_t)BLTZ_OUT, 1);
         return;
     }
 
@@ -86,12 +86,12 @@ void genbltz_out()
 void genbltz_idle()
 {
 #ifdef INTERPRET_BLTZ_IDLE
-    gencallinterp((uint32_t)BLTZ_IDLE, 1);
+    gencallinterp((uintptr_t)BLTZ_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZ_IDLE, 1);
+        gencallinterp((uintptr_t)BLTZ_IDLE, 1);
         return;
     }
 
@@ -138,12 +138,12 @@ void genbgez_test()
 void genbgez()
 {
 #ifdef INTERPRET_BGEZ
-    gencallinterp((uint32_t)BGEZ, 1);
+    gencallinterp((uintptr_t)BGEZ, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZ, 1);
+        gencallinterp((uintptr_t)BGEZ, 1);
         return;
     }
 
@@ -156,12 +156,12 @@ void genbgez()
 void genbgez_out()
 {
 #ifdef INTERPRET_BGEZ_OUT
-    gencallinterp((uint32_t)BGEZ_OUT, 1);
+    gencallinterp((uintptr_t)BGEZ_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZ_OUT, 1);
+        gencallinterp((uintptr_t)BGEZ_OUT, 1);
         return;
     }
 
@@ -174,12 +174,12 @@ void genbgez_out()
 void genbgez_idle()
 {
 #ifdef INTERPRET_BGEZ_IDLE
-    gencallinterp((uint32_t)BGEZ_IDLE, 1);
+    gencallinterp((uintptr_t)BGEZ_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZ_IDLE, 1);
+        gencallinterp((uintptr_t)BGEZ_IDLE, 1);
         return;
     }
 
@@ -192,12 +192,12 @@ void genbgez_idle()
 void genbltzl()
 {
 #ifdef INTERPRET_BLTZL
-    gencallinterp((uint32_t)BLTZL, 1);
+    gencallinterp((uintptr_t)BLTZL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZL, 1);
+        gencallinterp((uintptr_t)BLTZL, 1);
         return;
     }
 
@@ -210,12 +210,12 @@ void genbltzl()
 void genbltzl_out()
 {
 #ifdef INTERPRET_BLTZL_OUT
-    gencallinterp((uint32_t)BLTZL_OUT, 1);
+    gencallinterp((uintptr_t)BLTZL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZL_OUT, 1);
+        gencallinterp((uintptr_t)BLTZL_OUT, 1);
         return;
     }
 
@@ -228,12 +228,12 @@ void genbltzl_out()
 void genbltzl_idle()
 {
 #ifdef INTERPRET_BLTZL_IDLE
-    gencallinterp((uint32_t)BLTZL_IDLE, 1);
+    gencallinterp((uintptr_t)BLTZL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZL_IDLE, 1);
+        gencallinterp((uintptr_t)BLTZL_IDLE, 1);
         return;
     }
 
@@ -246,12 +246,12 @@ void genbltzl_idle()
 void genbgezl()
 {
 #ifdef INTERPRET_BGEZL
-    gencallinterp((uint32_t)BGEZL, 1);
+    gencallinterp((uintptr_t)BGEZL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZL, 1);
+        gencallinterp((uintptr_t)BGEZL, 1);
         return;
     }
 
@@ -264,12 +264,12 @@ void genbgezl()
 void genbgezl_out()
 {
 #ifdef INTERPRET_BGEZL_OUT
-    gencallinterp((uint32_t)BGEZL_OUT, 1);
+    gencallinterp((uintptr_t)BGEZL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZL_OUT, 1);
+        gencallinterp((uintptr_t)BGEZL_OUT, 1);
         return;
     }
 
@@ -282,12 +282,12 @@ void genbgezl_out()
 void genbgezl_idle()
 {
 #ifdef INTERPRET_BGEZL_IDLE
-    gencallinterp((uint32_t)BGEZL_IDLE, 1);
+    gencallinterp((uintptr_t)BGEZL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZL_IDLE, 1);
+        gencallinterp((uintptr_t)BGEZL_IDLE, 1);
         return;
     }
 
@@ -331,12 +331,12 @@ void genbranchlink()
 void genbltzal()
 {
 #ifdef INTERPRET_BLTZAL
-    gencallinterp((uint32_t)BLTZAL, 1);
+    gencallinterp((uintptr_t)BLTZAL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZAL, 1);
+        gencallinterp((uintptr_t)BLTZAL, 1);
         return;
     }
 
@@ -350,12 +350,12 @@ void genbltzal()
 void genbltzal_out()
 {
 #ifdef INTERPRET_BLTZAL_OUT
-    gencallinterp((uint32_t)BLTZAL_OUT, 1);
+    gencallinterp((uintptr_t)BLTZAL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZAL_OUT, 1);
+        gencallinterp((uintptr_t)BLTZAL_OUT, 1);
         return;
     }
 
@@ -369,12 +369,12 @@ void genbltzal_out()
 void genbltzal_idle()
 {
 #ifdef INTERPRET_BLTZAL_IDLE
-    gencallinterp((uint32_t)BLTZAL_IDLE, 1);
+    gencallinterp((uintptr_t)BLTZAL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZAL_IDLE, 1);
+        gencallinterp((uintptr_t)BLTZAL_IDLE, 1);
         return;
     }
 
@@ -388,12 +388,12 @@ void genbltzal_idle()
 void genbgezal()
 {
 #ifdef INTERPRET_BGEZAL
-    gencallinterp((uint32_t)BGEZAL, 1);
+    gencallinterp((uintptr_t)BGEZAL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZAL, 1);
+        gencallinterp((uintptr_t)BGEZAL, 1);
         return;
     }
 
@@ -407,12 +407,12 @@ void genbgezal()
 void genbgezal_out()
 {
 #ifdef INTERPRET_BGEZAL_OUT
-    gencallinterp((uint32_t)BGEZAL_OUT, 1);
+    gencallinterp((uintptr_t)BGEZAL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZAL_OUT, 1);
+        gencallinterp((uintptr_t)BGEZAL_OUT, 1);
         return;
     }
 
@@ -426,12 +426,12 @@ void genbgezal_out()
 void genbgezal_idle()
 {
 #ifdef INTERPRET_BGEZAL_IDLE
-    gencallinterp((uint32_t)BGEZAL_IDLE, 1);
+    gencallinterp((uintptr_t)BGEZAL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZAL_IDLE, 1);
+        gencallinterp((uintptr_t)BGEZAL_IDLE, 1);
         return;
     }
 
@@ -445,12 +445,12 @@ void genbgezal_idle()
 void genbltzall()
 {
 #ifdef INTERPRET_BLTZALL
-    gencallinterp((uint32_t)BLTZALL, 1);
+    gencallinterp((uintptr_t)BLTZALL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZALL, 1);
+        gencallinterp((uintptr_t)BLTZALL, 1);
         return;
     }
 
@@ -464,12 +464,12 @@ void genbltzall()
 void genbltzall_out()
 {
 #ifdef INTERPRET_BLTZALL_OUT
-    gencallinterp((uint32_t)BLTZALL_OUT, 1);
+    gencallinterp((uintptr_t)BLTZALL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZALL_OUT, 1);
+        gencallinterp((uintptr_t)BLTZALL_OUT, 1);
         return;
     }
 
@@ -483,12 +483,12 @@ void genbltzall_out()
 void genbltzall_idle()
 {
 #ifdef INTERPRET_BLTZALL_IDLE
-    gencallinterp((uint32_t)BLTZALL_IDLE, 1);
+    gencallinterp((uintptr_t)BLTZALL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BLTZALL_IDLE, 1);
+        gencallinterp((uintptr_t)BLTZALL_IDLE, 1);
         return;
     }
 
@@ -502,12 +502,12 @@ void genbltzall_idle()
 void genbgezall()
 {
 #ifdef INTERPRET_BGEZALL
-    gencallinterp((uint32_t)BGEZALL, 1);
+    gencallinterp((uintptr_t)BGEZALL, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZALL, 1);
+        gencallinterp((uintptr_t)BGEZALL, 1);
         return;
     }
 
@@ -521,12 +521,12 @@ void genbgezall()
 void genbgezall_out()
 {
 #ifdef INTERPRET_BGEZALL_OUT
-    gencallinterp((uint32_t)BGEZALL_OUT, 1);
+    gencallinterp((uintptr_t)BGEZALL_OUT, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZALL_OUT, 1);
+        gencallinterp((uintptr_t)BGEZALL_OUT, 1);
         return;
     }
 
@@ -540,12 +540,12 @@ void genbgezall_out()
 void genbgezall_idle()
 {
 #ifdef INTERPRET_BGEZALL_IDLE
-    gencallinterp((uint32_t)BGEZALL_IDLE, 1);
+    gencallinterp((uintptr_t)BGEZALL_IDLE, 1);
 #else
     if (((dst->addr & 0xFFF) == 0xFFC && (dst->addr < 0x80000000 || dst->addr >= 0xC0000000)) ||
         !g_core->cfg->is_compiled_jump_enabled)
     {
-        gencallinterp((uint32_t)BGEZALL_IDLE, 1);
+        gencallinterp((uintptr_t)BGEZALL_IDLE, 1);
         return;
     }
